bug: race condition
Fix a race condition where the configuration portion would set
some config values, then call execCC to set the final config 
value. If the component handler finished last the final value is
never set, but if it finished first, the configuration is never set to
"complete," barring functionality.

This commit ensures the component handler is a nice guy, and
additionally that it retrieves the final configuration value before
setting the config db.